+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+- Solution prop to RolePrivilege type
+
 # 0.7.3 (2020-07-23)
 
 ### Added